CloudFormation template to create a certificate in-region
For a regional CloudFront distribution, there must be a certificate that exists in the same region. The only certificate I had created was in us-east-1. I wanted the entire JSC Tracker stack to be in us-west-2, so I needed to create another certificate there.

The template says nothing about what region the certificate will be created in! This seems like a gap in CloudFormation itself. The way to specify the region is to give a --region argument to the 'aws cloudformation create-stack' command. But, this goes against the purpose of CloudFormation: to let you track everything that was done to create all the needed resources.