Define monkey-patches for additional Kernel methods
db - returns a reference to the DB singleton.

jsc_email_sha1_salt - returns a reference to the secret string used to salt the email hashing.

standard_json_handling - extracts the 'body' portion of the Lambda event, parses it as JSON, generates a standard error response if the JSON couldn't be parsed, otherwise invokes the supplied block. Standardizes the result which is normally a JSON-hash. Can return 200 for success, 404 if a Dynamo lookup fails, or 500 errors for anything else.