Properly support CORS with two possible Origins
Access-Control-Allow-Origin must be set in the response headers to match the caller's own Origin header in the request. This is, to me, a bit of an odd architecture. The server could just as well return a boolean for 'yes, fine' or 'no, not fine', but instead it returns the value back if it's acceptable. Anyway, it seems to make the browser happy.