Change User to use 'sub' as the pk, and drop the Google Auth logic
With this architecture, the browser (in JavaScript / React) displays the Login button, authenticates through Google, and sends the resulting access_token to the back-end. The back-end only has to verify the token and extract some vital statistics: the 'sub' or subject (a unique ID for the user), the email address, first and last name, and profile (a URL for the user's profile picture). That means a lot of the code for talking to Google is not needed on the back-end at all. We also can expect not to need to generate SHA1s of email addresses and I plan to drop that logic soon.