Define IdField and use it for #sub in User
When creating a new object, the ID field :sub must be supplied a value in the constructor. The pk is calculated from it.

When loading an existing object, the pk supplies a value for sub (it calculates it by removing the '_user' suffix).

IdField can only be given a value once by calling initialize_<field_name>.

Also drop some more references to 'label' and 'markdown' that were dropped.