get_user_user_id: Accept a dash instead of the actual ID
Common use case: The user wishes to view their own user information, and generally, the common information they are shown: information on their fellow JSC members. Passing a hyphen instead of the ID means, use the value of 'sub' from the Google Access Token as the user ID. If the user isn't found in DDB, create it immediately, using the values from the token.