Migrate contact_id_set to contact_id_list
It has become clear that we need this to be ordered; whenever a contact is modified, it should appear first in the list by default, and DynamoDB sets (rightly) do not promise to preserve ordering. Also, DynamoDB lists support being empty, whereas its sets do not (which is odd). We can remove the empty-set workaround if we're willing to do our own checking for duplicate IDs in the list.