Refactoring to define StandardJsonHandlerInput, and add check for /admin/ permission
If the path contains /admin/ then the current user must be an admin, or we have a problem.

The number of inputs to the standard_json_handler block rose to 5, and I decided it was time to factor out an object to hold these values.

I also noticed I was repeating the logic to look up users, and to treat '-' user_id as a synonym for the current user's sub. I factored it out and moved it into standard_json_handler. It also accepts an optional lambda to handle the case where the current user doesn't yet exist in the database and needs to be created; for now, this is only supplied from get_user_user_id.rb because it's expected that this will always be the first API hit by a new user after logging in.