Ensure that scalars are stored in DynamoDB as strings
Even integers are best stored as strings, so they aren't turned into floating points or rendered as scientific notation, etc. after saving and reloading.

The only collection-typed fields we have are lists and sets. So if the value is an Array or Set, leave it be. Otherwise it should always be safe to call #to_s on it.