Attempted refactor of post_admin_users_assign.rb to save writes until the end
I'd rather not read, write, read, write... especially the target JSC which, if we were given multiple users to move into it, would for sure be written N times for N users.

Instead, let's keep a Set containing all the objects (Users and Jscs) that were modified, then call #write! on all of them at the end.

Also, load all specified Users before we change anything. If we don't find one of them, call the whole thing off without modifying any data. Don't do a partial write.

In the event that two or more Users are coming _from_ the same JSC (also quite possible, if one JSC has shrunk and is being merged into another one), only load it once. Make repeated edits to the copy in memory and then write it once at the end.

Use the new User#unassigned?

Tighten up the definition of a legal JSC: it's a string of digits starting with a non-zero digit. A single regex match can verify this.