Sort the entries in jsc-tracker-api-gateway.yaml alphabetically by path
I did this by hand and may well have made a mistake. But the desired ordering goes like this:

* First some unchanging text at the start of the template (Parameters, the Resources line, JSCTrackerApi, domain name, and DNS entry)
* Then all the resources ordered by path, starting with / (the base path mapping) and alphabetically, so /admin comes before /user, and /user/z comes before /user/{a}
* Then all the main-method operations, sorted by path, and when there are two with the same path (as with GET /user/{user_id} and POST /user/{user_id}), alphabetically by the verb (GET before POST)
* Then all the OPTIONS entries, sorted by path
* Then JSCTrackerDeployment which is a unique entry, with the DependsOn: in the same ordering as the main-methods and OPTIONS entries appeared above it
* Then the permissions entries, sorted by path
* And finally the Outputs: section at the end which is fixed

In the process I noticed there was an OPTIONS /user which... I can't think of why I would need that. OPTIONS entries are for CORS and they should all correspond to real GET / POST operations, since they are only supposed to be called by the React components. There is never a need to hit the /user path without any arguments. I will try dropping it. Obviously I'll need to update the CloudFormation stack and test thoroughly after this is all done.